fix(search): address all 23 cycle-2 review findings for #355 lexical relevance
- Finding 2/13: fix stale AD-355-7 doc in SearchLayer trait (blast-radius
  path uses Some(K*limit).max(100) on every path, not None)
- Finding 3: consolidate CANDIDATE_POOL_FLOOR=100 + candidate_pool() helper
  shared by pure-lexical and blast-radius paths; compound K=4 keeps no floor
  (documented trade-off)
- Finding 6: document O(limit×K) clone cost in compound path as accepted for
  verify-then-truncate-LAST invariant (AD-355-2)
- Finding 8: remove stale sq.limit=None comment in blast-radius Step 1
- Finding 9/23: add NgramIndexReader::lexical_index_version() (reads 6 bytes:
  SKIX magic + FORMAT_VERSION); export LEXICAL_INDEX_FORMAT_VERSION from lib;
  check_staleness now checks lexical version before AST so v2 index triggers
  auto-rebuild instead of hard error (ADR-006 self-heal)
- Finding 10: add discriminating verify-gate test for compound path — lib.rs
  with trigram overlap but no literal must be dropped; auth.rs with literal
  must survive (PF-007)
- Finding 11: fix query_tests comment accuracy (tests trigram-miss path,
  not verify gate)
- Finding 12: add negative assertion to multi-file reader test (file 2 must
  be absent from raw trigram results)
- Finding 14: widen df to u64 before +1 in compute_idf (PF-004)
- Finding 15: query_substring_present returns false for empty/whitespace-only
  query (defense-in-depth); update tests and snippet doc comment accordingly
- Finding 16: replace "tracked in a follow-up" with "#356 (pool-K calibration)"
- Finding 17: add lexical n-gram index gotcha entry to CLAUDE.md
- Finding 20: large-file (>5MB) path now reads up to MAX_VERIFY_SCAN_BYTES
  for substring verification instead of returning verified=false unconditionally
- Finding 22: add rank assertion to single-file roundtrip reader test
- Staleness test suite: replace raw b"stub" writes for index.skidx with
  write_lexical_index_stub() helper (SKIX\x03\x00); add new discriminating
  test for lexical self-heal (v2 index → NoStoredHead + manifest returned)
- 2985 rskim tests pass; rskim-search nextest passes; clippy -D warnings clean
  on rskim, rskim-search, rskim-research; Snyk: 0 issues

+++ crates/rskim-research/src/idf.rs
@@ -17,13 +17,22 @@ use crate::types::{BigramWeight, TrigramWeight};
 ///
 /// Panics in debug builds if `total_docs == 0`, which would produce `NEG_INFINITY`.
 /// Callers must ensure the corpus is non-empty before invoking this function.
+///
+/// # PF-004 (widen before arithmetic)
+///
+/// `df` is widened to `u64` before adding 1 to avoid a wrapping overflow when
+/// `df == u32::MAX` (which would produce 0 as the denominator, yielding `+inf`).
+/// In practice `df` is bounded by the number of indexed files, so u32::MAX is
+/// unreachable — but the widen-before-arithmetic pattern is applied verbatim per
+/// PF-004 for correctness and consistency with the rest of the codebase.
 #[must_use]
 pub fn compute_idf(df: u32, total_docs: u32) -> f32 {
     debug_assert!(
         total_docs > 0,
         "total_docs must be > 0; got 0 — caller must guard against empty corpus"
     );
-    ((total_docs as f64) / ((df + 1) as f64)).ln() as f32 + 1.0
+    // PF-004: widen to u64 before adding 1 to prevent wrapping when df == u32::MAX.
+    ((total_docs as f64) / ((u64::from(df) + 1) as f64)).ln() as f32 + 1.0
 }
 
 /// Build the weight table from a document-frequency map.
